changed some structure to make it more readable
